Add folder structure and change routing
Add folder stucture for each format.  Currently every file is in the content folder, but if there are 65 files it might be more useful to split these up initially by the formats they belong to.  This also removed the need for a prefix describing the format it belongs to on each file.

The routing in express has been changed to reflect this new folder structure.  Now the content folder is searched for subfolders, and these subfolders are added to the file path for data to be read in.  The route still displays these as part of content, as it probably isn't useful for this to be explicit in navigation.