WIP: Add rake file for taxon data
Add rake file to pull in taxon data from the content store.  Files have reverted back to having the url data in the name with underscores as the formats this information is required for getting the taxon data from the content-store api, and the formats themselves do not contain this information.  E.g. there are two types of statistics, statics in publications and statistics announcements.  There are also other types of publications which are not statistics which have a different path.

Currently the rake file brings in metadata associated with each slug, and also the parents for each slug.  The taxon parents are not being used currently as work on this is in progress in the taxon team and needs to be finshed before grandparents etc are available under this section.  However once this work is complete we should switch over to using taxon parents as this is more current and should give a more accurate picture of taxonomy parents a page has.