Add examples of each format to index page
Not all of these are currently rendering correctly, and
not all of them fall under the same part of the taxonomy (some are early
years, some are KS1&2).

This is just to make it easy to work with the prototype, because some of
the old links pointed to things that are not actually in the prototype
anymore.

We may follow up on this by automatically generating links by format
https://trello.com/c/t85TTDmy/115-add-examples-of-formats-to-the-prototype-index