Recent documents should be recent
The ordering was failing because the sort callback was returning a
boolean instead of negative, 0, positive.

getTime() is elapsed time since 1970, and the difference in that
value does the right thing.

Also added a quick hack for manual sections - they don't include
their own timestamps. We might want to exclude formats
like manual section and html publication from all of these lists.