Replace content HTML
Clear out `app/content` and refetch everything from the Key Stage 1 & 2
inventory.

We need to make sure all the HTML is consistent with the JSON data
we fetch from content store and rummager, so that there are no dead
links on the topic pages and sidebars.

This excludes pages that are now redirected, because these need to be
handled carefully to ensure all the API calls work. It's probably
easier to resolve these once and then feed it through the import
pipeline again.