Add "more like this" links to all content pages
This commit adds functionality to query Rummager for more like this
results and returns those to the view, so we can display them.

It also uses a new data file that includes the content ids of the
taxons. This is because the taxonomy data file currently doesn't include
them and would be very time consuming to add them there.

We should revisit that decision when we are sure we are going to keep
related links.

For now, this is being added so information architects can see what
links are being recommended.