Link to accordion page
If a page is tagged to a topic that is more than 3 levels deep then it will be in an accordion. The breadcrumb and side nav should link to the accordion page and not the invididual topic page.

TO-FIX: This is bodged for the sidebar nav and will link to the first topic for every topic if the breadcrumb is longer that 3 items.