Use clearer terminology when determining partials
On the live site, we use the concept of Grandchildren to determine if we
should display a grid view or an accordion view. If a taxon has
grandchildren, we display a grid; otherwise we display an accordion.
When no Children or Grandchildren exist, we only show tagged content.

This commit aims at bringing that terminology into the prototype as
well.