Refactor taxons and taxon presenter
We recently dropped the reliance on a hard-coded JSON file that provided
our taxonomy. As a result, taxons are now fetched asynchronously, which
led to a restructuring where they rely heavily on promises.

This refactor focuses primarily on tidying up the taxon class, making
it responsible for fetching its own data asynchronously, and returning
a promise that will resolve once all the requested information has been
gathered.

This refactoring also affects the taxon presenter (which is obviously
tightly coupled to the taxon class) as well as a couple of other classes
that had previously used the taxon class.

Views have been updated to use the new taxon API, but this mostly
involved renaming properties.